requirements.txt
pillow
PyQT5
